Fix sign extension issue with clang/gcc
This is an adaptation of antonm's CL with mods suggested by asiva

http://codereview.chromium.org/9003015

Dartium Mac tests pass with this.

Review URL: http://codereview.chromium.org//9091016

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@2972 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/snapshot.cc
@@ -257,7 +257,7 @@ void SnapshotWriter::WriteObject(RawObject* rawobj) {
 
   // First check if it is a Smi (i.e not a heap object).
   if (!rawobj->IsHeapObject()) {
-    Write<int64_t>(reinterpret_cast<int64_t>(rawobj));
+    Write<int64_t>(reinterpret_cast<intptr_t>(rawobj));
     return;
   }
